requirements.txt
flask 
flask_sqlalchemy
python-dotenv
flask_wtf
wtforms
smtplib
